Workhorse for simulation studies
Source: R/evalGrids.R

Generates data according to all provided constellations in dataGrid and applies all provided constellations in procGrid to them.

evalGrids(dataGrid, procGrid = expandGrid(proc = "length"),
  replications = 1, discardGeneratedData = FALSE, progress = FALSE,
  summary.fun = NULL, ncpus = 1L, cluster = NULL,
  clusterSeed = rep(12345, 6), clusterLibraries = NULL,
  clusterGlobalObjects = NULL, fallback = NULL, envir = globalenv(), ...)

Arguments

dataGrid
a data.frame where the first column is a character vector with function names. The other columns contain parameters for the functions specified in the first column. Parameters with NA are ignored.

procGrid
similar as dataGrid the first column must contain function names. The other columns contain parameters for the functions specified in the first column. The data generated according to dataGrid will always be passed to the first unspecified argument of the functions specified in the first column of procGrid.

replications
number of replications for the simulation

discardGeneratedData
if TRUE the generated data is deleted after all function constellations in procGrid have been applied. Otherwise, ALL generated data sets will be part of the returned object.

progress
if TRUE a progress bar is shown in the console.

summary.fun
univariate functions to summarize the results (numeric or logical) over the replications, e.g. mean, sd. Alternatively, summary.fun can be one function that may return a vector.

ncpus
a cluster of ncpus workers (R-processes) is created on the local machine to conduct the simulation. If ncpus equals one no cluster is created and the simulation is conducted by the current R-process.

cluster
a cluster generated by the parallel package that will be used to conduct the simulation. If cluster is specified, then ncpus will be ignored.

clusterSeed
if the simulation is done in parallel manner, then the combined multiple-recursive generator from L'Ecuyer (1999) is used to generate random numbers. Thus clusterSeed must be a (signed) integer vector of length 6. The 6 elements of the seed are internally regarded as 32-bit unsigned integers. Neither the first three nor the last three should be all zero, and they are limited to less than 4294967087 and 4294944443 respectively.

clusterLibraries
a character vector specifying the packages that should be loaded by the workers.

clusterGlobalObjects
a character vector specifying the names of R objects in the global environment that should be exported to the global environment of every worker.

fallback
must be missing or a character specifying a file. Every time when the data generation function is changed, the results so far obtained are saved in the file specified by fallback.

envir
must be provided if the functions specified in dataGrid or procGrid are not part of the global environment.

...
only needed to alert the user if some deprecated arguments were used.

Value

The returned object is a list of the class evalGrid, where the fourth element is a list of lists named simulation. simulation[[i]][[r]] contains:

data

the data set that was generated by the ith constellation (ith row) of dataGrid in the rth replication

results

a list containing nrow(procGrid) objects. The jth object is the returned value of the function specified by the jth constellation (jth row) of procGrid applied to the data set contained in data

Note

If cluster is provided by the user the function evalGrids will NOT stop the cluster. This has to be done by the user. Conducting parallel simulations by specifying ncpus will internally create a cluster and stop it after the simulation is done.

See also

as.data.frame.evalGrid

Examples

rng = function(data, ...) {
ret = range(data)
names(ret) = c("min", "max")
ret
}

# call runif(n=1), runif(n=2), runif(n=3)
# and range on the three "datasets"
# generated by runif(n=1), runif(n=2), runif(n=3)
eg = evalGrids(
 expandGrid(fun="runif", n=1:3),
 expandGrid(proc="rng"),
 rep=10
)

#> Warning: The function evalGrids() is deprecated. Please use eval_tibbles() instead.

#> Warning: The function expandGrid() is deprecated. Please use expand_tibble() instead.

#> Warning: The function expandGrid() is deprecated. Please use expand_tibble() instead.

#> [1] "Estimated replications per hour:  61933939"

eg

#> $call
#> evalGrids(dataGrid = expandGrid(fun = "runif", n = 1:3), procGrid = expandGrid(proc = "rng"), 
#>     replications = 10)
#> 
#> $dataGrid
#>     fun n
#> 1 runif 1
#> 2 runif 2
#> 3 runif 3
#> 
#> $procGrid
#>   proc
#> 1  rng
#> 
#> $simulation
#> $simulation[[1]]
#> $simulation[[1]][[1]]
#> $simulation[[1]][[1]]$data
#> [1] 0.0888955
#> 
#> $simulation[[1]][[1]]$results
#> $simulation[[1]][[1]]$results[[1]]
#>       min       max 
#> 0.0888955 0.0888955 
#> 
#> 
#> 
#> $simulation[[1]][[2]]
#> $simulation[[1]][[2]]$data
#> [1] 0.527871
#> 
#> $simulation[[1]][[2]]$results
#> $simulation[[1]][[2]]$results[[1]]
#>      min      max 
#> 0.527871 0.527871 
#> 
#> 
#> 
#> $simulation[[1]][[3]]
#> $simulation[[1]][[3]]$data
#> [1] 0.2349702
#> 
#> $simulation[[1]][[3]]$results
#> $simulation[[1]][[3]]$results[[1]]
#>       min       max 
#> 0.2349702 0.2349702 
#> 
#> 
#> 
#> $simulation[[1]][[4]]
#> $simulation[[1]][[4]]$data
#> [1] 0.2585438
#> 
#> $simulation[[1]][[4]]$results
#> $simulation[[1]][[4]]$results[[1]]
#>       min       max 
#> 0.2585438 0.2585438 
#> 
#> 
#> 
#> $simulation[[1]][[5]]
#> $simulation[[1]][[5]]$data
#> [1] 0.3336488
#> 
#> $simulation[[1]][[5]]$results
#> $simulation[[1]][[5]]$results[[1]]
#>       min       max 
#> 0.3336488 0.3336488 
#> 
#> 
#> 
#> $simulation[[1]][[6]]
#> $simulation[[1]][[6]]$data
#> [1] 0.1119666
#> 
#> $simulation[[1]][[6]]$results
#> $simulation[[1]][[6]]$results[[1]]
#>       min       max 
#> 0.1119666 0.1119666 
#> 
#> 
#> 
#> $simulation[[1]][[7]]
#> $simulation[[1]][[7]]$data
#> [1] 0.03739304
#> 
#> $simulation[[1]][[7]]$results
#> $simulation[[1]][[7]]$results[[1]]
#>        min        max 
#> 0.03739304 0.03739304 
#> 
#> 
#> 
#> $simulation[[1]][[8]]
#> $simulation[[1]][[8]]$data
#> [1] 0.05453934
#> 
#> $simulation[[1]][[8]]$results
#> $simulation[[1]][[8]]$results[[1]]
#>        min        max 
#> 0.05453934 0.05453934 
#> 
#> 
#> 
#> $simulation[[1]][[9]]
#> $simulation[[1]][[9]]$data
#> [1] 0.2673686
#> 
#> $simulation[[1]][[9]]$results
#> $simulation[[1]][[9]]$results[[1]]
#>       min       max 
#> 0.2673686 0.2673686 
#> 
#> 
#> 
#> $simulation[[1]][[10]]
#> $simulation[[1]][[10]]$data
#> [1] 0.4276272
#> 
#> $simulation[[1]][[10]]$results
#> $simulation[[1]][[10]]$results[[1]]
#>       min       max 
#> 0.4276272 0.4276272 
#> 
#> 
#> 
#> 
#> $simulation[[2]]
#> $simulation[[2]][[1]]
#> $simulation[[2]][[1]]$data
#> [1] 0.09517665 0.70176252
#> 
#> $simulation[[2]][[1]]$results
#> $simulation[[2]][[1]]$results[[1]]
#>        min        max 
#> 0.09517665 0.70176252 
#> 
#> 
#> 
#> $simulation[[2]][[2]]
#> $simulation[[2]][[2]]$data
#> [1] 0.5477511 0.9548898
#> 
#> $simulation[[2]][[2]]$results
#> $simulation[[2]][[2]]$results[[1]]
#>       min       max 
#> 0.5477511 0.9548898 
#> 
#> 
#> 
#> $simulation[[2]][[3]]
#> $simulation[[2]][[3]]$data
#> [1] 0.07946089 0.62555272
#> 
#> $simulation[[2]][[3]]$results
#> $simulation[[2]][[3]]$results[[1]]
#>        min        max 
#> 0.07946089 0.62555272 
#> 
#> 
#> 
#> $simulation[[2]][[4]]
#> $simulation[[2]][[4]]$data
#> [1] 0.1484082 0.0372463
#> 
#> $simulation[[2]][[4]]$results
#> $simulation[[2]][[4]]$results[[1]]
#>       min       max 
#> 0.0372463 0.1484082 
#> 
#> 
#> 
#> $simulation[[2]][[5]]
#> $simulation[[2]][[5]]$data
#> [1] 0.9696468 0.1724914
#> 
#> $simulation[[2]][[5]]$results
#> $simulation[[2]][[5]]$results[[1]]
#>       min       max 
#> 0.1724914 0.9696468 
#> 
#> 
#> 
#> $simulation[[2]][[6]]
#> $simulation[[2]][[6]]$data
#> [1] 0.4399818 0.8340476
#> 
#> $simulation[[2]][[6]]$results
#> $simulation[[2]][[6]]$results[[1]]
#>       min       max 
#> 0.4399818 0.8340476 
#> 
#> 
#> 
#> $simulation[[2]][[7]]
#> $simulation[[2]][[7]]$data
#> [1] 0.43535228 0.09026915
#> 
#> $simulation[[2]][[7]]$results
#> $simulation[[2]][[7]]$results[[1]]
#>        min        max 
#> 0.09026915 0.43535228 
#> 
#> 
#> 
#> $simulation[[2]][[8]]
#> $simulation[[2]][[8]]$data
#> [1] 0.7563909 0.2780298
#> 
#> $simulation[[2]][[8]]$results
#> $simulation[[2]][[8]]$results[[1]]
#>       min       max 
#> 0.2780298 0.7563909 
#> 
#> 
#> 
#> $simulation[[2]][[9]]
#> $simulation[[2]][[9]]$data
#> [1] 0.7310415 0.7597456
#> 
#> $simulation[[2]][[9]]$results
#> $simulation[[2]][[9]]$results[[1]]
#>       min       max 
#> 0.7310415 0.7597456 
#> 
#> 
#> 
#> $simulation[[2]][[10]]
#> $simulation[[2]][[10]]$data
#> [1] 0.2564088137 0.0003509724
#> 
#> $simulation[[2]][[10]]$results
#> $simulation[[2]][[10]]$results[[1]]
#>          min          max 
#> 0.0003509724 0.2564088137 
#> 
#> 
#> 
#> 
#> $simulation[[3]]
#> $simulation[[3]][[1]]
#> $simulation[[3]][[1]]$data
#> [1] 0.4282120 0.1318027 0.9594336
#> 
#> $simulation[[3]][[1]]$results
#> $simulation[[3]][[1]]$results[[1]]
#>       min       max 
#> 0.1318027 0.9594336 
#> 
#> 
#> 
#> $simulation[[3]][[2]]
#> $simulation[[3]][[2]]$data
#> [1] 0.40122899 0.61148999 0.04283093
#> 
#> $simulation[[3]][[2]]$results
#> $simulation[[3]][[2]]$results[[1]]
#>        min        max 
#> 0.04283093 0.61148999 
#> 
#> 
#> 
#> $simulation[[3]][[3]]
#> $simulation[[3]][[3]]$data
#> [1] 0.80651670 0.06773117 0.93108990
#> 
#> $simulation[[3]][[3]]$results
#> $simulation[[3]][[3]]$results[[1]]
#>        min        max 
#> 0.06773117 0.93108990 
#> 
#> 
#> 
#> $simulation[[3]][[4]]
#> $simulation[[3]][[4]]$data
#> [1] 0.4304459 0.2033120 0.8845817
#> 
#> $simulation[[3]][[4]]$results
#> $simulation[[3]][[4]]$results[[1]]
#>       min       max 
#> 0.2033120 0.8845817 
#> 
#> 
#> 
#> $simulation[[3]][[5]]
#> $simulation[[3]][[5]]$data
#> [1] 0.34129542 0.06753315 0.53957040
#> 
#> $simulation[[3]][[5]]$results
#> $simulation[[3]][[5]]$results[[1]]
#>        min        max 
#> 0.06753315 0.53957040 
#> 
#> 
#> 
#> $simulation[[3]][[6]]
#> $simulation[[3]][[6]]$data
#> [1] 0.8030478 0.2773124 0.9348883
#> 
#> $simulation[[3]][[6]]$results
#> $simulation[[3]][[6]]$results[[1]]
#>       min       max 
#> 0.2773124 0.9348883 
#> 
#> 
#> 
#> $simulation[[3]][[7]]
#> $simulation[[3]][[7]]$data
#> [1] 0.6890035 0.5900332 0.2133564
#> 
#> $simulation[[3]][[7]]$results
#> $simulation[[3]][[7]]$results[[1]]
#>       min       max 
#> 0.2133564 0.6890035 
#> 
#> 
#> 
#> $simulation[[3]][[8]]
#> $simulation[[3]][[8]]$data
#> [1] 0.9091252 0.5738601 0.7215302
#> 
#> $simulation[[3]][[8]]$results
#> $simulation[[3]][[8]]$results[[1]]
#>       min       max 
#> 0.5738601 0.9091252 
#> 
#> 
#> 
#> $simulation[[3]][[9]]
#> $simulation[[3]][[9]]$data
#> [1] 0.9472091 0.6944337 0.6712095
#> 
#> $simulation[[3]][[9]]$results
#> $simulation[[3]][[9]]$results[[1]]
#>       min       max 
#> 0.6712095 0.9472091 
#> 
#> 
#> 
#> $simulation[[3]][[10]]
#> $simulation[[3]][[10]]$data
#> [1] 0.2434018 0.3578725 0.7407287
#> 
#> $simulation[[3]][[10]]$results
#> $simulation[[3]][[10]]$results[[1]]
#>       min       max 
#> 0.2434018 0.7407287 
#> 
#> 
#> 
#> 
#> 
#> $summary.fun
#> NULL
#> 
#> $est.reps.per.hour
#> [1] 61933939
#> 
#> $sessionInfo
#> R version 3.4.4 (2018-03-15)
#> Platform: x86_64-pc-linux-gnu (64-bit)
#> Running under: Arch Linux
#> 
#> Matrix products: default
#> BLAS: /usr/lib/libblas.so.3.8.0
#> LAPACK: /usr/lib/liblapack.so.3.8.0
#> 
#> locale:
#>  [1] LC_CTYPE=en_US.UTF-8       LC_NUMERIC=C              
#>  [3] LC_TIME=en_US.UTF-8        LC_COLLATE=en_US.UTF-8    
#>  [5] LC_MONETARY=en_US.UTF-8    LC_MESSAGES=en_US.UTF-8   
#>  [7] LC_PAPER=en_US.UTF-8       LC_NAME=C                 
#>  [9] LC_ADDRESS=C               LC_TELEPHONE=C            
#> [11] LC_MEASUREMENT=en_US.UTF-8 LC_IDENTIFICATION=C       
#> 
#> attached base packages:
#> [1] stats     graphics  grDevices utils     datasets  methods   base     
#> 
#> other attached packages:
#> [1] simTool_1.1.0  testthat_2.0.0
#> 
#> loaded via a namespace (and not attached):
#>  [1] Rcpp_0.12.16       pillar_1.2.1       compiler_3.4.4     plyr_1.8.4        
#>  [5] bindr_0.1          tools_3.4.4        highlight_0.4.7.2  pkgbuild_1.0.0    
#>  [9] pkgload_1.0.0      digest_0.6.15      debugme_1.1.0      memoise_1.1.0     
#> [13] evaluate_0.10.1    tibble_1.4.2       pkgconfig_2.0.1    rlang_0.2.0       
#> [17] cli_1.0.0          rstudioapi_0.7     commonmark_1.4     parallel_3.4.4    
#> [21] curl_3.1           yaml_2.1.18        pkgdown_0.1.0.9000 bindrcpp_0.2      
#> [25] withr_2.1.2        dplyr_0.7.4        stringr_1.3.0      httr_1.3.1        
#> [29] roxygen2_6.0.1     xml2_1.2.0         knitr_1.20         desc_1.1.1        
#> [33] fs_1.2.2           rprojroot_1.3-2    reshape_0.8.7      glue_1.2.0        
#> [37] R6_2.2.2           rematch_1.0.1      rmarkdown_1.9      callr_2.0.2       
#> [41] whisker_0.3-2      tidyr_0.8.0        purrr_0.2.4        magrittr_1.5      
#> [45] backports_1.1.2    htmltools_0.3.6    MASS_7.3-49        assertthat_0.2.0  
#> [49] stringi_1.1.6      crayon_1.3.4      
#> 
#> attr(,"class")
#> [1] "evalGrid"

# summarizing the results in a data.frame
as.data.frame(eg)

#> Warning: The function as.data.frame.evalGrid() is deprecated. Please use eval_tibbles, since it always returns a tibble-object.

#> Warning: The function expandGrid() is deprecated. Please use expand_tibble() instead.

#>    i j   fun n proc replication          min        max
#> 1  1 1 runif 1  rng           1 0.0888955038 0.08889550
#> 2  1 1 runif 1  rng           2 0.5278710033 0.52787100
#> 3  1 1 runif 1  rng           3 0.2349702110 0.23497021
#> 4  1 1 runif 1  rng           4 0.2585437574 0.25854376
#> 5  1 1 runif 1  rng           5 0.3336487842 0.33364878
#> 6  1 1 runif 1  rng           6 0.1119666023 0.11196660
#> 7  1 1 runif 1  rng           7 0.0373930435 0.03739304
#> 8  1 1 runif 1  rng           8 0.0545393399 0.05453934
#> 9  1 1 runif 1  rng           9 0.2673685643 0.26736856
#> 10 1 1 runif 1  rng          10 0.4276272098 0.42762721
#> 11 2 1 runif 2  rng           1 0.0951766476 0.70176252
#> 12 2 1 runif 2  rng           2 0.5477511259 0.95488981
#> 13 2 1 runif 2  rng           3 0.0794608853 0.62555272
#> 14 2 1 runif 2  rng           4 0.0372463040 0.14840821
#> 15 2 1 runif 2  rng           5 0.1724914415 0.96964677
#> 16 2 1 runif 2  rng           6 0.4399818102 0.83404762
#> 17 2 1 runif 2  rng           7 0.0902691513 0.43535228
#> 18 2 1 runif 2  rng           8 0.2780297908 0.75639087
#> 19 2 1 runif 2  rng           9 0.7310414694 0.75974559
#> 20 2 1 runif 2  rng          10 0.0003509724 0.25640881
#> 21 3 1 runif 3  rng           1 0.1318026731 0.95943357
#> 22 3 1 runif 3  rng           2 0.0428309273 0.61148999
#> 23 3 1 runif 3  rng           3 0.0677311674 0.93108990
#> 24 3 1 runif 3  rng           4 0.2033120390 0.88458166
#> 25 3 1 runif 3  rng           5 0.0675331461 0.53957040
#> 26 3 1 runif 3  rng           6 0.2773123942 0.93488835
#> 27 3 1 runif 3  rng           7 0.2133564144 0.68900347
#> 28 3 1 runif 3  rng           8 0.5738601099 0.90912522
#> 29 3 1 runif 3  rng           9 0.6712094849 0.94720912
#> 30 3 1 runif 3  rng          10 0.2434017900 0.74072871

# we now generate data for a regression
# and fit different regression models

# not that we use SD and not sd (the
# reason for this is the cast() call below)
regData = function(n, SD){
 data.frame(
   x=seq(0,1,length=n),
   y=rnorm(n, sd=SD))
}

eg = evalGrids(
 expandGrid(fun="regData", n=20, SD=1:2),
 expandGrid(proc="lm", formula=c("y~x", "y~I(x^2)")),
 replications=2)

#> Warning: The function evalGrids() is deprecated. Please use eval_tibbles() instead.

#> Warning: The function expandGrid() is deprecated. Please use expand_tibble() instead.

#> Warning: The function expandGrid() is deprecated. Please use expand_tibble() instead.

#> [1] "Estimated replications per hour:  682555"

# can not be converted to data.frame, because
# an object of class "lm" can not converted to
# a data.frame
try(as.data.frame(eg))

#> Warning: The function as.data.frame.evalGrid() is deprecated. Please use eval_tibbles, since it always returns a tibble-object.

#> Warning: The function expandGrid() is deprecated. Please use expand_tibble() instead.

# for the data.frame we just extract the r.squared
# from the fitted model
as.data.frame(eg, convert.result.fun=function(fit) c(rsq=summary(fit)$r.squared))

#> Warning: The function as.data.frame.evalGrid() is deprecated. Please use eval_tibbles, since it always returns a tibble-object.

#> Warning: The function expandGrid() is deprecated. Please use expand_tibble() instead.

#>   i j     fun  n SD proc  formula replication        rsq
#> 1 1 1 regData 20  1   lm      y~x           1 0.04033998
#> 2 1 1 regData 20  1   lm      y~x           2 0.08286291
#> 3 1 2 regData 20  1   lm y~I(x^2)           1 0.04082008
#> 4 1 2 regData 20  1   lm y~I(x^2)           2 0.07815913
#> 5 2 1 regData 20  2   lm      y~x           1 0.02829111
#> 6 2 1 regData 20  2   lm      y~x           2 0.09343087
#> 7 2 2 regData 20  2   lm y~I(x^2)           1 0.02687260
#> 8 2 2 regData 20  2   lm y~I(x^2)           2 0.03008225

# for the data.frame we just extract the coefficients
# from the fitted model
df = as.data.frame(eg, convert.result.fun=coef)

#> Warning: The function as.data.frame.evalGrid() is deprecated. Please use eval_tibbles, since it always returns a tibble-object.

#> Warning: The function expandGrid() is deprecated. Please use expand_tibble() instead.

# since we have done 2 replication we can calculate
# sum summary statistics
library("reshape")

#> 
#> Attaching package: ‘reshape’

#> The following objects are masked from ‘package:simTool’:
#> 
#>     rename, round_any

df$replication=NULL
mdf = melt(df, id=1:7, na.rm=TRUE)
cast(mdf, ... ~ ., c(mean, length, sd))

#>   i j     fun  n SD proc  formula    variable        mean length         sd
#> 1 1 1 regData 20  1   lm      y~x (Intercept) -0.04840345      2 0.05979441
#> 2 1 1 regData 20  1   lm      y~x           x  0.81537381      2 0.21827869
#> 3 1 2 regData 20  1   lm y~I(x^2) (Intercept)  0.09397553      2 0.01530314
#> 4 1 2 regData 20  1   lm y~I(x^2)      I(x^2)  0.77551549      2 0.18897131
#> 5 2 1 regData 20  2   lm      y~x (Intercept) -0.58444666      2 0.26059373
#> 6 2 1 regData 20  2   lm      y~x           x  1.50407915      2 0.54672630
#> 7 2 2 regData 20  2   lm y~I(x^2) (Intercept) -0.18938794      2 0.53776108
#> 8 2 2 regData 20  2   lm y~I(x^2)      I(x^2)  1.04348249      2 0.01111998

# note if the data.frame would contain the column
# named "sd" instead of "SD" the cast will generate
# an error
names(df)[5] = "sd"
mdf = melt(df, id=1:7, na.rm=TRUE)
try(cast(mdf, ... ~ ., c(mean, length, sd)))

# extracting the summary of the fitted.model
as.data.frame(eg, convert.result.fun=function(x) {
 ret = coef(summary(x))
 data.frame(valueName = rownames(ret), ret, check.names=FALSE)
})

#> Warning: The function as.data.frame.evalGrid() is deprecated. Please use eval_tibbles, since it always returns a tibble-object.

#> Warning: The function expandGrid() is deprecated. Please use expand_tibble() instead.

#>    i j     fun  n SD proc  formula replication   valueName     Estimate
#> 1  1 1 regData 20  1   lm      y~x           1 (Intercept) -0.006122416
#> 2  1 1 regData 20  1   lm      y~x           1           x  0.661027467
#> 3  1 1 regData 20  1   lm      y~x           2 (Intercept) -0.090684476
#> 4  1 1 regData 20  1   lm      y~x           2           x  0.969720150
#> 5  1 2 regData 20  1   lm y~I(x^2)           1 (Intercept)  0.104796482
#> 6  1 2 regData 20  1   lm y~I(x^2)           1      I(x^2)  0.641892597
#> 7  1 2 regData 20  1   lm y~I(x^2)           2 (Intercept)  0.083154573
#> 8  1 2 regData 20  1   lm y~I(x^2)           2      I(x^2)  0.909138384
#> 9  2 1 regData 20  2   lm      y~x           1 (Intercept) -0.768714257
#> 10 2 1 regData 20  2   lm      y~x           1           x  1.117485281
#> 11 2 1 regData 20  2   lm      y~x           2 (Intercept) -0.400179071
#> 12 2 1 regData 20  2   lm      y~x           2           x  1.890673026
#> 13 2 2 regData 20  2   lm y~I(x^2)           1 (Intercept) -0.569642445
#> 14 2 2 regData 20  2   lm y~I(x^2)           1      I(x^2)  1.051345499
#> 15 2 2 regData 20  2   lm y~I(x^2)           2 (Intercept)  0.190866568
#> 16 2 2 regData 20  2   lm y~I(x^2)           2      I(x^2)  1.035619479
#>    Std. Error     t value  Pr(>|t|)
#> 1   0.4444818 -0.01377428 0.9891616
#> 2   0.7599309  0.86985206 0.3958332
#> 3   0.4447614 -0.20389466 0.8407233
#> 4   0.7604089  1.27526138 0.2184311
#> 5   0.3407551  0.30754192 0.7619609
#> 6   0.7333972  0.87523188 0.3929772
#> 7   0.3419283  0.24319304 0.8106057
#> 8   0.7359221  1.23537310 0.2325736
#> 9   0.9028769 -0.85140540 0.4057291
#> 10  1.5436494  0.72392429 0.4784161
#> 11  0.8119239 -0.49287755 0.6280578
#> 12  1.3881470  1.36201213 0.1899972
#> 13  0.6928550 -0.82216690 0.4217398
#> 14  1.4912114  0.70502780 0.4898163
#> 15  0.6439904  0.29638109 0.7703315
#> 16  1.3860415  0.74717783 0.4646055

# we now compare to methods for
# calculating quantiles

# the functions and parameters
# that generate the data
N = c(10, 50, 100)
library("plyr")

#> 
#> Attaching package: ‘plyr’

#> The following objects are masked from ‘package:reshape’:
#> 
#>     rename, round_any

dg = rbind.fill(
 expandGrid(fun="rbeta", n=N, shape1=4, shape2=4),
 expandGrid(fun="rnorm", n=N))

#> Warning: The function expandGrid() is deprecated. Please use expand_tibble() instead.

#> Warning: The function expandGrid() is deprecated. Please use expand_tibble() instead.

# definition of the two quantile methods
emp.q = function(data, probs) c(quantile(data, probs=probs))
nor.q = function(data, probs) {
 ret = qnorm(probs, mean=mean(data), sd=sd(data))
 names(ret) = names(quantile(1, probs=probs))
 ret
}

# the functions and parameters that are
# applied to the generate data
pg = rbind.fill(expandGrid(proc=c("emp.q", "nor.q"), probs=c(0.01, 0.025, 0.05)))

#> Warning: The function expandGrid() is deprecated. Please use expand_tibble() instead.

# generate data and apply quantile methods
set.seed(1234)
eg = evalGrids(dg, pg, replication=50, progress=TRUE)

#> Warning: The function evalGrids() is deprecated. Please use eval_tibbles() instead.

#> Error in get(procGrid[j, 1], envir = envir): object 'emp.q' not found

# convert the results to a data.frame
df = as.data.frame(eg)

#> Warning: The function as.data.frame.evalGrid() is deprecated. Please use eval_tibbles, since it always returns a tibble-object.

#> Warning: The function expandGrid() is deprecated. Please use expand_tibble() instead.

#> Error in list_to_dataframe(res, attr(.data, "split_labels"), .id, id_as_factor): Results must be all atomic, or all data frames

df$replication=NULL
mdf = melt(df, id=1:8, na.rm=TRUE)

# calculate, print and plot summary statistics
require("ggplot2")

#> Loading required package: ggplot2

print(a <- arrange(cast(mdf, ... ~ ., c(mean, sd)), n))

#> Error in `[.data.frame`(data, , vars, drop = FALSE): undefined columns selected

ggplot(a, aes(x=fun, y=mean, color=proc)) + geom_point(size=I(3)) + facet_grid(probs ~ n)

#> Error in ggplot(a, aes(x = fun, y = mean, color = proc)): object 'a' not found