Conduct Simulation Studies with a Minimal Amount of Source Code

The simTool package is designed for statistical simulations that have two components. One component generates the data and the other one analyzes the data. The main aims of the simTool package are the reduction of the administrative source code (mainly loops and management code for the results) and a simple applicability of the package that allows the user to quickly learn how to work with the simTool package. Parallel computing is also supported. Finally, convenient functions are provided to summarize the simulation results.

Details

Package:	simTool

Type:	Package

Version:	1.0.3

Date:	2013-02-21

License:	GPL-3

Package:

evalGrids is the workhorse. as.data.frame is function coercing the result object of evalGrids to a data.frame. expandGrid is only a convenient function

Examples

dg = expand_tibble(fun="rexp", n=c(10, 20), rate=1:2)
pg = expand_tibble(proc="summary")
eval_tibbles(dg, pg, replications=3)

#> # A tibble: 72 x 6
#>    fun       n  rate replications proc    results
#>    <chr> <dbl> <int>        <int> <chr>     <dbl>
#>  1 rexp     10     1            1 summary 0.0167 
#>  2 rexp     10     1            1 summary 0.133  
#>  3 rexp     10     1            1 summary 0.372  
#>  4 rexp     10     1            1 summary 0.523  
#>  5 rexp     10     1            1 summary 0.936  
#>  6 rexp     10     1            1 summary 1.29   
#>  7 rexp     10     1            2 summary 0.00217
#>  8 rexp     10     1            2 summary 0.240  
#>  9 rexp     10     1            2 summary 0.445  
#> 10 rexp     10     1            2 summary 0.763  
#> # ... with 62 more rows
#> Number of data generating functions: 4
#> Number of analyzing procedures: 1
#> Number of replications: 3
#> Estimated replications per hour: 3287024
#> Start of the simulation: 2018-10-13 17:33:53
#> End of the simulation: 2018-10-13 17:33:53

eval_tibbles(dg, pg, replications=3, summary_fun = list(mean = mean))

#> # A tibble: 4 x 12
#>   fun       n  rate replications summary_fun proc      Min. `1st Qu.` Median
#>   <chr> <dbl> <int>        <int> <chr>       <chr>    <dbl>     <dbl>  <dbl>
#> 1 rexp     10     1            1 mean        summary 0.235      0.612  0.877
#> 2 rexp     20     1            1 mean        summary 0.0474     0.379  0.753
#> 3 rexp     10     2            1 mean        summary 0.0452     0.207  0.380
#> 4 rexp     20     2            1 mean        summary 0.0366     0.244  0.436
#> # ... with 3 more variables: Mean <dbl>, `3rd Qu.` <dbl>, Max. <dbl>
#> Number of data generating functions: 4
#> Number of analyzing procedures: 1
#> Number of replications: 3
#> Estimated replications per hour: 583623
#> Start of the simulation: 2018-10-13 17:33:53
#> End of the simulation: 2018-10-13 17:33:53

eval_tibbles(dg, pg, replications=3, summary_fun = list(mean = mean, sd = sd))

#> # A tibble: 8 x 12
#>   fun       n  rate replications summary_fun proc      Min. `1st Qu.` Median
#>   <chr> <dbl> <int>        <int> <chr>       <chr>    <dbl>     <dbl>  <dbl>
#> 1 rexp     10     1            1 mean        summary 0.0269   0.239    0.494
#> 2 rexp     10     1            1 sd          summary 0.0168   0.0843   0.127
#> 3 rexp     20     1            1 mean        summary 0.0720   0.404    0.905
#> 4 rexp     20     1            1 sd          summary 0.0514   0.222    0.464
#> 5 rexp     10     2            1 mean        summary 0.120    0.248    0.441
#> 6 rexp     10     2            1 sd          summary 0.0543   0.00730  0.134
#> 7 rexp     20     2            1 mean        summary 0.0145   0.216    0.454
#> 8 rexp     20     2            1 sd          summary 0.0123   0.0734   0.125
#> # ... with 3 more variables: Mean <dbl>, `3rd Qu.` <dbl>, Max. <dbl>
#> Number of data generating functions: 4
#> Number of analyzing procedures: 1
#> Number of replications: 3
#> Estimated replications per hour: 457366
#> Start of the simulation: 2018-10-13 17:33:54
#> End of the simulation: 2018-10-13 17:33:54