Workhorse for simulation studies
Source: R/eval_tibbles.R

Generates data according to all provided constellations in data_tibble and applies all provided constellations in proc_tibble to them.

eval_tibbles(data_grid, proc_grid = expand_tibble(proc = "length"),
  replications = 1, discard_generated_data = FALSE,
  post_analyze = identity, summary_fun = NULL, group_for_summary = NULL,
  ncpus = 1L, cluster = NULL, cluster_seed = rep(12345, 6),
  cluster_libraries = NULL, cluster_global_objects = NULL,
  envir = globalenv(), simplify = TRUE)

Arguments

data_grid
a data.frame or tibble where the first column is a character vector with function names. The other columns contain parameters for the functions specified in the first column. Parameters with NA are ignored.

proc_grid
similar as data_grid the first column must contain function names. The other columns contain parameters for the functions specified in the first column. The data generated according to data_grid will always be passed to the first unspecified argument of the functions specified in the first column of proc_grid.

replications
number of replications for the simulation

discard_generated_data
if TRUE the generated data is deleted after all function constellations in proc_grid have been applied. Otherwise, ALL generated data sets will be part of the returned object.

post_analyze
this is a convenience function, that is applied directly after the data analyzing function.

summary_fun
named list of univariate function to summarize the results (numeric or logical) over the replications, e.g. list(mean = mean, sd = sd).

group_for_summary
if the result returned by the data analyzing function or post_analyze is a data.frame with more than one row, one usually is interested in summarizing the results while grouping for some variables. This group variables can be passed as a character vector into group_for_summary

ncpus
a cluster of ncpus workers (R-processes) is created on the local machine to conduct the simulation. If ncpus equals one no cluster is created and the simulation is conducted by the current R-process.

cluster
a cluster generated by the parallel package that will be used to conduct the simulation. If cluster is specified, then ncpus will be ignored.

cluster_seed
if the simulation is done in parallel manner, then the combined multiple-recursive generator from L'Ecuyer (1999) is used to generate random numbers. Thus cluster_seed must be a (signed) integer vector of length 6. The 6 elements of the seed are internally regarded as 32-bit unsigned integers. Neither the first three nor the last three should be all zero, and they are limited to less than 4294967087 and 4294944443 respectively.

cluster_libraries
a character vector specifying the packages that should be loaded by the workers.

cluster_global_objects
a character vector specifying the names of R objects in the global environment that should be exported to the global environment of every worker.

envir
must be provided if the functions specified in data_grid or proc_grid are not part of the global environment.

simplify
usually the result column is nested, by default it is tried to unnest it.

Value

The returned object list of the class eval_tibbles, where the element simulations contain the results of the simulation.

Note

If cluster is provided by the user the function eval_tibbles will NOT stop the cluster. This has to be done by the user. Conducting parallel simulations by specifying ncpus will internally create a cluster and stop it after the simulation is done.

Examples

rng = function(data, ...) {
ret = range(data)
names(ret) = c("min", "max")
ret
}

dg <- expand_tibble(fun = "rnorm", n = c(5L, 10L))
pg <- expand_tibble(proc = c("rng", "median", "length"))

eval_tibbles(dg, pg,rep = 2)

#>   |                                                                              |                                                                      |   0%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================================================| 100%

#> # A tibble: 16 x 5
#>    fun       n replications proc   results
#>    <chr> <int>        <int> <chr>    <dbl>
#>  1 rnorm     5            1 rng     -1.78 
#>  2 rnorm     5            1 rng     -0.430
#>  3 rnorm     5            1 median  -0.723
#>  4 rnorm     5            1 length   5.00 
#>  5 rnorm     5            2 rng     -1.41 
#>  6 rnorm     5            2 rng      1.88 
#>  7 rnorm     5            2 median  -1.04 
#>  8 rnorm     5            2 length   5.00 
#>  9 rnorm    10            1 rng     -0.654
#> 10 rnorm    10            1 rng      1.74 
#> 11 rnorm    10            1 median   0.450
#> 12 rnorm    10            1 length  10.0  
#> 13 rnorm    10            2 rng     -0.830
#> 14 rnorm    10            2 rng      1.62 
#> 15 rnorm    10            2 median  -0.132
#> 16 rnorm    10            2 length  10.0  
#> Number of data generating functions: 2
#> Number of analyzing procedures: 3
#> Number of replications: 2
#> Estimated replications per hour: 285119
#> Start of the simulation: 2018-03-25 20:33:22
#> End of the simulation: 2018-03-25 20:33:22

eval_tibbles(dg, pg,rep = 2, post_analyze = purrr::compose(tibble::as_tibble, t))

#>   |                                                                              |                                                                      |   0%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================================================| 100%

#> # A tibble: 12 x 7
#>    fun       n replications proc      min    max      V1
#>    <chr> <int>        <int> <chr>   <dbl>  <dbl>   <dbl>
#>  1 rnorm     5            1 rng     -1.14  1.39   NA    
#>  2 rnorm     5            1 median  NA    NA       0.132
#>  3 rnorm     5            1 length  NA    NA       5.00 
#>  4 rnorm     5            2 rng     -1.05  0.803  NA    
#>  5 rnorm     5            2 median  NA    NA      -0.360
#>  6 rnorm     5            2 length  NA    NA       5.00 
#>  7 rnorm    10            1 rng     -1.30  2.23   NA    
#>  8 rnorm    10            1 median  NA    NA       0.955
#>  9 rnorm    10            1 length  NA    NA      10.0  
#> 10 rnorm    10            2 rng     -1.83  1.85   NA    
#> 11 rnorm    10            2 median  NA    NA       0.366
#> 12 rnorm    10            2 length  NA    NA      10.0  
#> Number of data generating functions: 2
#> Number of analyzing procedures: 3
#> Number of replications: 2
#> Estimated replications per hour: 1620030
#> Start of the simulation: 2018-03-25 20:33:22
#> End of the simulation: 2018-03-25 20:33:22

eval_tibbles(dg, pg,rep = 2, summary_fun = list(mean = mean, sd = sd))

#>   |                                                                              |                                                                      |   0%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================================================| 100%

#> # A tibble: 12 x 8
#>    fun       n replications summary_fun proc        min    max    value
#>    <chr> <int>        <int> <chr>       <chr>     <dbl>  <dbl>    <dbl>
#>  1 rnorm     5            1 mean        rng     -0.476   1.81   NA     
#>  2 rnorm     5            1 mean        median  NA      NA       0.517 
#>  3 rnorm     5            1 mean        length  NA      NA       5.00  
#>  4 rnorm     5            1 sd          rng      0.356   0.662  NA     
#>  5 rnorm     5            1 sd          median  NA      NA       0.0412
#>  6 rnorm     5            1 sd          length  NA      NA       0.    
#>  7 rnorm    10            1 mean        rng     -1.40    1.71   NA     
#>  8 rnorm    10            1 mean        median  NA      NA      -0.326 
#>  9 rnorm    10            1 mean        length  NA      NA      10.0   
#> 10 rnorm    10            1 sd          rng      0.0774  0.849  NA     
#> 11 rnorm    10            1 sd          median  NA      NA       0.365 
#> 12 rnorm    10            1 sd          length  NA      NA       0.    
#> Number of data generating functions: 2
#> Number of analyzing procedures: 3
#> Number of replications: 2
#> Estimated replications per hour: 488167
#> Start of the simulation: 2018-03-25 20:33:22
#> End of the simulation: 2018-03-25 20:33:22

regData = function(n, SD){
  data.frame(
  x=seq(0,1,length=n),
  y=rnorm(n, sd=SD))
}

eg <- eval_tibbles(
  expand_tibble(fun="regData", n=5L, SD=1:2),
  expand_tibble(proc="lm", formula=c("y~x", "y~I(x^2)")),
  group_for_summary = "term",
  replications=3
)

#>   |                                                                              |                                                                      |   0%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================================================| 100%

eg

#> # A tibble: 12 x 7
#>    fun         n    SD replications proc  formula  results 
#>    <chr>   <int> <int>        <int> <chr> <chr>    <list>  
#>  1 regData     5     1            1 lm    y~x      <S3: lm>
#>  2 regData     5     1            1 lm    y~I(x^2) <S3: lm>
#>  3 regData     5     1            2 lm    y~x      <S3: lm>
#>  4 regData     5     1            2 lm    y~I(x^2) <S3: lm>
#>  5 regData     5     1            3 lm    y~x      <S3: lm>
#>  6 regData     5     1            3 lm    y~I(x^2) <S3: lm>
#>  7 regData     5     2            1 lm    y~x      <S3: lm>
#>  8 regData     5     2            1 lm    y~I(x^2) <S3: lm>
#>  9 regData     5     2            2 lm    y~x      <S3: lm>
#> 10 regData     5     2            2 lm    y~I(x^2) <S3: lm>
#> 11 regData     5     2            3 lm    y~x      <S3: lm>
#> 12 regData     5     2            3 lm    y~I(x^2) <S3: lm>
#> Number of data generating functions: 2
#> Number of analyzing procedures: 2
#> Number of replications: 3
#> Estimated replications per hour: 641821
#> Start of the simulation: 2018-03-25 20:33:22
#> End of the simulation: 2018-03-25 20:33:22

presever_rownames = function(mat)
{
  rn = rownames(mat)
  ret = tibble::as_tibble(mat)
  ret$term = rn
  ret
}

eg <- eval_tibbles(
  expand_tibble(fun="regData", n=5L, SD=1:2),
  expand_tibble(proc="lm", formula=c("y~x", "y~I(x^2)")),
  post_analyze = purrr::compose(presever_rownames, coef, summary),
  #post_analyze = broom::tidy, # is a nice out of the box alternative
  summary_fun = list(mean = mean, sd = sd),
  group_for_summary = "term",
  replications=3
)

#>   |                                                                              |                                                                      |   0%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================================================| 100%

tidyr::unnest(eg$simulation)

#> # A tibble: 16 x 12
#>    fun         n    SD replications summary_fun proc  formula term     Estimate
#>    <chr>   <int> <int>        <int> <chr>       <chr> <chr>   <chr>       <dbl>
#>  1 regData     5     1            1 mean        lm    y~x     (Interc…   0.207 
#>  2 regData     5     1            1 mean        lm    y~x     x         -0.541 
#>  3 regData     5     1            1 mean        lm    y~I(x^… (Interc…  -0.0268
#>  4 regData     5     1            1 mean        lm    y~I(x^… I(x^2)    -0.0969
#>  5 regData     5     1            1 sd          lm    y~x     (Interc…   0.707 
#>  6 regData     5     1            1 sd          lm    y~x     x          1.20  
#>  7 regData     5     1            1 sd          lm    y~I(x^… (Interc…   0.769 
#>  8 regData     5     1            1 sd          lm    y~I(x^… I(x^2)     1.38  
#>  9 regData     5     2            1 mean        lm    y~x     (Interc…  -0.726 
#> 10 regData     5     2            1 mean        lm    y~x     x          0.988 
#> 11 regData     5     2            1 mean        lm    y~I(x^… (Interc…  -0.583 
#> 12 regData     5     2            1 mean        lm    y~I(x^… I(x^2)     0.936 
#> 13 regData     5     2            1 sd          lm    y~x     (Interc…   0.131 
#> 14 regData     5     2            1 sd          lm    y~x     x          4.18  
#> 15 regData     5     2            1 sd          lm    y~I(x^… (Interc…   0.995 
#> 16 regData     5     2            1 sd          lm    y~I(x^… I(x^2)     2.92  
#> # ... with 3 more variables: `Std. Error` <dbl>, `t value` <dbl>,
#> #   `Pr(>|t|)` <dbl>